pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rahatline-py"
version = "0.1.0"
authors = [
    {name="Sepehr", email="rahatbin.com@gmail.com"},
]
description = "Create connection for rahatline"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "pyjwt==2.3.0",
    "websockets==15.0.1",
    "aiortc==1.13.0"
]

[project.urls]
Homepage = "https://github.com/yourusername/my-library"
Source = "https://github.com/yourusername/my-library"

[tool.setuptools.packages.find]
where = ["src"]     